Delete AppCompatCacheParser.exe
This is the comment from the author of AppCompatCache, Eric Zimmerman. We are going to remove it for now. 

"This is a false positive on something. i have recompiled and tried several ways to stop this, and only this, program from hitting on AV. i have no idea why. the source is on github as well. for some reasons the scanners are hitting on the packing or the kinds of access the tool is doing reading into the Registry. Not much i can do unfortunately but i can assure you it is not trojaned or otherwise as all my software is built on the same machine and nothing else has hit on AV."