Updates for APT Hunter
For APT hunter, we leveraged other researchers's tools / scripts to
parse the information.  We want to specially say thanks to Eric
Zimmerman, author of AmCache Parser and AppCompatCacheParser, Mark Woan,
the author of ShimCacheParser, Harlan Carvey, author of all HC tools,
and Joakim Schicht,  author of RAWCOPY.

We know that he AppCompatCacheParser by Eric has been recently flagged
by AV due to the fact that some AV vendors incorrectly detected the
obfuscation and packing tool, Smart Assembly.  We will try to work on
this matter. For now, we are still using the new AppCompatCacheParser ,
version 0.9.0.0, released by Eric recently , the updated version can be
used to  look for all available ControlSet keys and processes them all
for AppCompatCache values found in a hive.

Josh and I are still working on the PowerShell version of APT Hunter,
and it will be released soon